v4.0

+++ entity/user/User.go
@@ -9,6 +9,8 @@ import (
 	"io/ioutil"
 	"os"
 	"path/filepath"
+
+	"github.com/tpisntgod/Agenda/entity/mylog"
 )
 
 var userItemsFilePath string = "src/github.com/tpisntgod/Agenda/Json/UserItems.json"
@@ -33,49 +35,6 @@ func init() {
 	CurrentUser = nil
 	readJSON()
 }
-
-// 测试部分
-/*
-func main() {
-	//IsLogInOutTest()
-	RegisterUserTest()
-	//IsRegisteredUserTest()
-	//DeleteUserTest()
-}
-
-func IsLogInOutTest() {
-	fmt.Println("Test IsLogin():")
-	fmt.Println(IsLogin())
-	LoginUser("huziang", "123456")
-	fmt.Println(IsLogin())
-	LogoutUser()
-	fmt.Println(IsLogin())
-}
-func RegisterUserTest() {
-	fmt.Println("Test RegisterUser():")
-	err := RegisterUser("huziang", "123456", "3254266353@qq.com", "13719316539")
-	fmt.Println(err)
-	err = RegisterUser("houhongxiao", "123645", "325423@qq.com", "12321331213")
-	fmt.Println(err)
-	LoginUser("huziang", "123456")
-	ListUsers()
-}
-
-func IsRegisteredUserTest() {
-	fmt.Println("Test IsRegisteredUser():")
-	fmt.Println(IsRegisteredUser("houhongxiao"))
-}
-
-func DeleteUserTest() {
-	err := LoginUser("huziang", "123456")
-	fmt.Println(err)
-	DeleteUser()
-	err = LoginUser("houhongxiao", "123645")
-	fmt.Println(err)
-	ListUsers()
-	LogoutUser()
-}
-*/
 
 // 新建一个userItem，并返回指针
 func newUser(name string, password string,
@@ -118,6 +77,7 @@ func RegisterUser(name string, password string,
 	userItems[name] = *newUser(name, password, email, phoneNumber)
 
 	writeJSON()
+	mylog.AddLog("", "RegisterUser", "", userItems[name].String())
 	return nil
 }
 
@@ -143,6 +103,7 @@ func LoginUser(name string, password string) error {
 	// 成功登录
 	CurrentUser = &tempUser
 	writeJSON()
+	mylog.AddLog(GetLogonUsername(), "LoginUser", "", "")
 	fmt.Println("Hi " + name + "!")
 	return nil
 }
@@ -155,6 +116,7 @@ func LogoutUser() error {
 
 	CurrentUser = nil
 	writeJSON()
+	mylog.AddLog(GetLogonUsername(), "LogoutUser", "", "")
 	fmt.Println("Logout successfully!")
 	return nil
 }
@@ -182,6 +144,9 @@ func ListUsers() error {
 	// 输出结尾
 	outputStr += "All user listed as follow.\n"
 	fmt.Printf("%s", outputStr)
+
+	mylog.AddLog(GetLogonUsername(), "ListUsers", "", "")
+
 	return nil
 }
 
@@ -193,6 +158,9 @@ func DeleteUser() error {
 	}
 
 	delete(userItems, CurrentUser.Name)
+
+	mylog.AddLog(GetLogonUsername(), "DeleteUser", (*CurrentUser).String(), "")
+
 	CurrentUser = nil
 	writeJSON()
 	return nil
@@ -239,8 +207,11 @@ func writeJSON() {
 	}
 
 	// 写入CurrentUser
+
 	if CurrentUser == nil {
-		os.Remove(currentUserFilePath)
+		if _, err := os.Open(currentUserFilePath); err == nil {
+			os.Remove(currentUserFilePath)
+		}
 		return
 	}
 	b2, err2 := json.Marshal(*CurrentUser)
@@ -250,4 +221,9 @@ func writeJSON() {
 		}
 		ioutil.WriteFile(currentUserFilePath, b2, 0755)
 	}
+}
+
+// to string
+func (u userItem) String() string {
+	return "{Name:" + u.Name + "  Email:" + u.Email + "  Phone:" + u.PhoneNumber + "}"
 }
